--- a/config/profile.conf
+++ b/config/profile.conf
@@ -2,11 +2,11 @@
     Type: "consumer.Profiler"
     Enable: true
     Runs: 100000
-    Batches: 1000
+    Batches: 100
     Characters: "abcdefghijklmnopqrstuvwxyz .,!;:-_"
     Message: "%256s"
     Streams: "profile"
-    KeepRunning: true
+    KeepRunning: false
 
 "Benchmark":
     Type: "producer.Benchmark"