--- a/core/messagehelper_test.go
+++ b/core/messagehelper_test.go
@@ -53,7 +53,7 @@ func TestGetAppliedContentFunction(t *testing.T) {
 func TestGetAppliedContentFromPayload(t *testing.T) {
 	expect := ttesting.NewExpect(t)
 	resultFunc := GetAppliedContentGetFunction("payload")
-	msg := NewMessage(nil, []byte("message payload"), 1)
+	msg := NewMessage(nil, []byte("message payload"), nil, 1)
 
 	expect.Equal("message payload", string(resultFunc(msg)))
 }
@@ -61,7 +61,7 @@ func TestGetAppliedContentFromPayload(t *testing.T) {
 func TestGetAppliedContentFromMetadata(t *testing.T) {
 	expect := ttesting.NewExpect(t)
 	resultFunc := GetAppliedContentGetFunction("foo")
-	msg := NewMessage(nil, []byte("message payload"), 1)
+	msg := NewMessage(nil, []byte("message payload"), nil, 1)
 	msg.GetMetadata().SetValue("foo", []byte("foo content"))
 
 	expect.Equal("foo content", string(resultFunc(msg)))