--- a/core/metrics.go
+++ b/core/metrics.go
@@ -17,6 +17,7 @@ package core
 import (
 	"fmt"
 	"github.com/trivago/tgo"
+	"sync"
 	"time"
 )