--- a/format/streamroute.go
+++ b/format/streamroute.go
@@ -71,7 +71,7 @@ func (format *StreamRoute) ApplyFormatter(msg *core.Message) error {
 
 	default:
 		streamName := content[:delimiterIdx]
-		streamMsg := core.NewMessage(nil, []byte(streamName), msg.GetStreamID())
+		streamMsg := core.NewMessage(nil, []byte(streamName), nil, msg.GetStreamID())
 
 		content = content[(delimiterIdx + len(format.delimiter)):]
 		format.SetAppliedContent(msg, content)