--- a/glide.lock
+++ b/glide.lock
@@ -1,5 +1,5 @@
 hash: c86ab655c5028444ab0002cd5d7a7a0c9f6a1498474fb580cde5fafbf98103a8
-updated: 2017-07-03T10:00:53.674320348+02:00
+updated: 2017-07-03T17:30:57.015268357+02:00
 imports:
 - name: github.com/abbot/go-http-auth
   version: efc9484eee77263a11f158ef4f30fcc30298a942
@@ -118,7 +118,7 @@ imports:
 - name: github.com/sirupsen/logrus
   version: 202f25545ea4cf9b191ff7f846df5d87c9382c2b
 - name: github.com/trivago/tgo
-  version: c787d684e0a8d84d70a3199b01262c80b8afaf3a
+  version: 0c9f1c59c165c624611a767f2383fe2bf679c07d
   subpackages:
   - tcontainer
   - terrors
@@ -134,7 +134,7 @@ imports:
 - name: github.com/x-cray/logrus-prefixed-formatter
   version: 0b2c791a02f18788b281dc30e243a0632dad695b
 - name: golang.org/x/crypto
-  version: 5ef0053f77724838734b6945dd364d3847e5de1d
+  version: 5746f0a2e262b80ccdf839a4a37a6b97962210f1
   subpackages:
   - bcrypt
   - blowfish