Improving general pipeline performance (config/profile.conf) by ~38%

Before:
```
Overview: 10000000 messages sent in 15.8775 seconds
Avg: 629823 msg/sec
Best: 0.1484 sec = 673646 msg/sec
Worst: 0.2193 sec = 455968 msg/sec
```

After:
```
Overview: 10000000 messages sent in 11.4712 seconds
Avg: 871746 msg/sec
Best: 0.1064 sec = 939957 msg/sec
Worst: 0.1732 sec = 577518 msg/sec
```

Optimisations include
* Added the possibility to create messages with existing metadata
* Lock reduction and speed improvements of stream based metrics
* Lock reduction of stream registry
* Pre-pooling of byte slices used by message cloning/creation