@arnecls 2017-07-04
Also fixing #96